Test mode for RED certification
Implements "START_SOFTAP" command. The command switches the module
into softap mode running silently (no beacons) on a given channel.
SSID of the AP is "_test". The command accepts single u8 parameter
which is interpreted as the number of the channel the AP listens on.
If the channel number is less than 100 the AP broadcasts garbage
frames continuously on that channel. If the parameter value is grater
than 100, the AP just listens silently on channel number value - 100.